+++ webClient/src/app/shared/dialog/open-dataset/open-dataset.component.ts
@@ -16,7 +16,7 @@ import { ENDPOINTS } from '../../../../environments/environment';
 @Component({
   selector: 'app-open-dataset',
   templateUrl: './open-dataset.component.html',
-  styleUrls: ['./open-dataset.component.scss']
+  styleUrls: ['../../../../styles.scss']
 })
 export class OpenDatasetComponent implements OnInit {