+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Zlux Editor Changelog
 
+## `2.9.0`
+- Added the feature to update the dataset in the editor
+- Added the check for e-tag while updating the dataset
+- Added feature to open a file in new broswer tab
+- Added the ability to view the dataset write conflicts in the diff-viewer
+- Added the ability to overwrite the dataset and accept incoming change inside the diff-viewer
+
 ## `2.8.0`
+- Bugfix: Fixed error message & phantom tab when opening undefined length dataset
+- Bugfix: Unable to recalculate the size of monaco-code-container when switching from diff-viewer to code-editor.
+- Added previously selected content comparison (Diff viewer)
 - Added the ability to refresh the file properties in File Tree, when an already exiting untagged file is saved with an encoding type.
 - Added the ability to use the latest USS encoding when saving an exiting file.
+- Removed the Cancel button and replaced it with X button on all dialogue pop-ups
+- Added toggle bar to show and hide File Explorer.
 
 ## `2.7.0`
 - Added function to toggle the file explorer